DestinationHint and Angel bugfix 2 BoundingBox improved

+++ controltheland/src/main/java/com/btxtech/game/services/bot/impl/BotSyncBaseItem.java
@@ -97,7 +97,7 @@ public class BotSyncBaseItem {
         idle = true;
     }
 
-    public int getDistanceTo(SyncBaseItem syncBaseItem) {
+    public double getDistanceTo(SyncBaseItem syncBaseItem) {
         return this.syncBaseItem.getSyncItemArea().getDistance(syncBaseItem);
     }
 }
